requirements.txt

coverage==6.2
Django==3.0.5
django-rest-swagger==2.2.0
djangorestframework==3.12.1
djongo==1.3.3
pymongo==3.12.1
pytest-bdd
pytest-django